un-breakup v2_steps.py so it can be easier to see changes
